update the Vagrantfile for recent changes
No longer able to pass the commit as a parameter. It could be set as an
environment variable, but given /etc/puppet is mapped to the current
directory, it doesn't serve any purpose. Allow the NODE_TYPE to be altered by
an environment variable to test a provision with master_base, for example